@@ -2,11 +2,11 @@
   "name": "jde-framework",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^12.0.0",
-    "@angular/core": "^12.0.0",
+    "@angular/common": "^19.1.2",
+    "@angular/core": "^19.1.2",
     "jde-material": "0.0.1"
   },
   "dependencies": {
-    "tslib": "^2.1.0"
+    "tslib": "^2.8.1"
   }
 }
