@@ -1,41 +1,30 @@
-import {Component, Inject, Input, Output, AfterViewInit, EventEmitter, ViewChild, ViewChildren, ElementRef, OnInit, OnDestroy, QueryList, ChangeDetectorRef} from '@angular/core';
-import { MatFormField } from '@angular/material/form-field';
+import {Component, effect, Inject, input, output, AfterViewInit, EventEmitter, ViewChild, ViewChildren, ElementRef, OnInit, OnDestroy, QueryList, ChangeDetectorRef, computed, viewChildren, model, signal, CUSTOM_ELEMENTS_SCHEMA} from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MatFormFieldModule, MatLabel } from '@angular/material/form-field';
 import { ActivatedRoute, Router } from '@angular/router';
 import { IErrorService } from '../../../services/error/IErrorService';
 import { ComponentPageTitle } from 'jde-material';
-import { Field, FieldKind, IEnum, IGraphQL, IQueryResult } from '../../../services/IGraphQL';
+import { Field, FieldKind, IEnum, IGraphQL, TableSchema } from 'jde-framework';
 import { StringUtils } from '../../../utilities/StringUtils';
+import { MatSelectModule } from '@angular/material/select';
+import { MatInputModule } from '@angular/material/input';
+import { MatChipGrid, MatChipsModule } from '@angular/material/chips';
+import { MatButtonModule } from '@angular/material/button';
 
-@Component( { selector: 'graph-ql-properties', templateUrl: 'properties.html'} )
-export class GraphQLProperties implements OnInit, AfterViewInit, OnDestroy{
-	constructor( private route: ActivatedRoute, private router:Router, private componentPageTitle:ComponentPageTitle, @Inject('IGraphQL') private graphQL: IGraphQL, private cdr: ChangeDetectorRef, @Inject('IErrorService') private cnsl: IErrorService )
-	{}
-	async ngOnDestroy(){
-		for( const f of this.fields.filter(f=>f.type==InputTypes.Select) ){
-			if( this.#original[f.name] )
-				this.#original[f.name] = f.options.find( (x)=>x.id==this.#original[f.name] ).name;
-		}
+@Component({
+    selector: 'properties',
+    templateUrl: 'properties.html',
+    imports: [CommonModule, MatButtonModule, MatChipsModule, MatChipGrid, MatInputModule, MatFormFieldModule, MatLabel, MatSelectModule],
+		schemas: [CUSTOM_ELEMENTS_SCHEMA]
+})
+export class Properties implements OnInit{
+	constructor( private route: ActivatedRoute, private router:Router, private componentPageTitle:ComponentPageTitle, @Inject('IGraphQL') private graphQL: IGraphQL, private cdr: ChangeDetectorRef, @Inject('IErrorService') private cnsl: IErrorService ){
+		effect( ()=>{
+			this.componentPageTitle.detail = this.record()["name"] ?? `New ${this.schema().type}`;
+		});
 	}
+
 	async ngOnInit(){
-		try{
-			const schema = await this.graphQL.schema( [this.type] );
-			for( const field of schema[0].fields.filter((x)=>[FieldKind.OBJECT,FieldKind.LIST].indexOf(x.type.underlyingKind)==-1 && GraphQLProperties.noShowFields.indexOf(x.name)==-1) ){
-				let values:Array<IEnum>;
-				if( field.type.underlyingKind==FieldKind.ENUM ){
-					values = ( await this.graphQL.query<IQueryResult<IEnum>>(` __type(name: "${field.type.name}") { enumValues { id name } }`) ).__type["enumValues"];
-					if( this.#original[field.name] ){
-						const v = values.find( (x)=>x.name==this.#original[field.name] )?.id;
-						this.clone.set( field.name, v );
-						this.#original[field.name] = v;
-					}
-				}
-				const p = new PropertyField(field, values);
-				(p.type==InputTypes.Bool ? this.boolFields : this.fields).push( p );
-			}
-		}
-		catch( e ){
-			console.log( e );
-		}
 		const order = ["target", "name"];
 		const sort = ( x:PropertyField,y:PropertyField )=>{
 			const yIndex = order.indexOf( y.name )+1;
@@ -45,96 +34,56 @@ export class GraphQLProperties implements OnInit, AfterViewInit, OnDestroy{
 			else
 				return x.name.localeCompare( y.name );
 		}
-		this.fields.sort( sort );
+		this.fields().sort( sort );
 
-		this.viewPromise = Promise.resolve( true );
-		setTimeout( ()=>{ this.viewChildren.first["nativeElement"].focus(); this.cdr.detectChanges();}, 0 );
+		this.isLoading.set( false );
 	}
 
-	@ViewChildren('myInput') viewChildren:QueryList<MatFormField>;
-
-	ngAfterViewInit()
-	{}
+	onChange( field:string, value:string ){
+		let f = this.ctor();
+		let newRecord = new f( this.record() );
+		newRecord[field] = value;
+		this.record.set( newRecord );
+	}
 
 	originalOrder = ( a, b )=> {return 0;}
-	onCancelClick(){
-		this.router.navigate( ['..'], { relativeTo: this.route } );
+
+	fields = computed<PropertyField[]>( ()=>{
+		let y = [];
+		let filter = (field)=>
+			[FieldKind.OBJECT,FieldKind.LIST,FieldKind.LIST].indexOf(field.type.underlyingKind)==-1
+			&& field.type.ofType?.name!='Boolean'
+			&& Properties.noShowFields.indexOf(field.name)==-1
+			&& this.excludedColumns().indexOf(field.name)==-1;
+		for( const field of this.schema().fields.filter(filter) ){
+			let values = field.type.underlyingKind==FieldKind.ENUM ? this.schema().enums.get(field.type.name) : undefined;
+			y.push( new PropertyField(field, values) );
+		}
+		return y;
+	});
+	boolFields = computed<PropertyField[]>( ()=>{
+		return [];
+		//return this.fields().filter( (x)=>x.type==InputTypes.Bool );
+	});
+	getEnumId( field:PropertyField ):number{
+		let stringValue = this.record()[field.name];
+		return stringValue ? field.options.find( (x)=>x.name==stringValue ).id : 0;
 	}
 
-	mods():any{
-		let input:any = this.original.id==null ? {...this.clone} : {};
-		for( var m in this.original ){
-			if( this.clone.get(m)!==undefined && this.original[m]!=this.clone.get(m) )
-				input[m] = this.clone.get(m);
-		}
+	ctor = input.required<new (item: any) => any>();
+	excludedColumns = input<string[]>([]);
+	record = model.required<any>();
+	schema = input.required<TableSchema>();
+	type = input.required<string>();
 
-		for( var [key,value] of this.clone ){//previously no description, now description.
-			if( this.original[key]===undefined )
-				input[key] = this.clone.get(key);
-		}
-		for( var field of this.boolFields){
-			if( this.clone.get(field.name)===undefined )
-				input[field.name] = false;
-		}
-		return input;
-	}
+	stringFields = viewChildren<ElementRef>( "stringField" );
 
-	get enableSubmit():boolean{
-		let enable = !this.saving
-			&& this.fields.find( (x)=>!x.nullable && x.type!=InputTypes.Select && (!this.clone.get(x.name) || this.clone.get(x.name).length==0) )==null; //non-null is null
-		if( enable && this.original.id!=null )
-			enable = Object.keys( this.mods() ).length>0;
-		return enable;
-	}
+	isLoading = signal<boolean>( true );
 
-	async onSubmitClick(){
-		const update = this.original.id!=null;
-		let idString = update ? `"id":${this.original.id},` : "";
-		let output = update ? "" : "{id}";
-		let cmd = update ? "update" : "create";
-		const input = this.mods();
-		if( Object.keys(input).length ){
-			var ql = `${cmd}${this.type}( ${idString} "input": ${JSON.stringify(input)} )${output}`;
-			try{
-				await this.graphQL.mutation( ql );
-				this.router.navigate( ['..'], { relativeTo: this.route } );
-				// if( update && this.original.target==this.clone["target"] )
-				// {
-				// 	for( let m in input )
-				// 		this.original[m] = input[m];
-				// 	this.save.emit( this.original );
-				// }
-				// else
-				// 	this.router.navigate([`../${input.target ?? this.original.target}`], { relativeTo: this.route });
-			}
-			catch( e ){
-				this.cnsl.error( "Saving properties failed", e );
-				console.log( e.toString() );
-			}
-			this.saving=false;
-		}
-	}
 	get InputTypes(){ return InputTypes; }
-	fields=new Array<PropertyField>();
-	boolFields=new Array<PropertyField>();
-	saving=false;
-	@Output() save = new EventEmitter<any>();
 	static noShowFields = ["id", "created", "attributes", "updated", "deleted"];
-	clone = new Map<string,any>();//good but tabs don't work.
-	@Input() set original(x){
-		this.#original = x ?? {};
-		this.componentPageTitle.detail = this.#original["name"] ?? "New";
-		//let add = ( m )=>this.clone.set( m, this.#original[m] ?? '' );
-		for( let m in x ){
-			if( x[m]!==undefined && GraphQLProperties.noShowFields.indexOf(m)==-1 )
-				this.clone.set( m, x[m] );
-		}
-	} get original(){return this.#original; } #original:any;
-	@Input() type:string;
-	viewPromise:Promise<boolean>;
 }
-enum InputTypes
-{
+enum InputTypes{
 	Select=-1,
 	None=0,
 	Text=1,
