@@ -1,7 +1,5 @@
-//import { Observable } from 'rxjs';
 
-export interface IProfile
-{
+export interface IProfile{
 	get<T>( key:string ):Promise<T>;
 	put<T>( name:string, value:T ):void;
 	putJson( name:string, json:string ):void;
