@@ -1,14 +1,15 @@
 import { webSocket, WebSocketSubject } from 'rxjs/webSocket';
-import { firstValueFrom, Observable, Subject } from 'rxjs';
-import { HttpClient, HttpErrorResponse, HttpResponse } from '@angular/common/http';
-import { Table, Mutation } from './IGraphQL';
+import { firstValueFrom } from 'rxjs';
+import { HttpClient, HttpResponse } from '@angular/common/http';
+import { FieldKind, fromIsoDuration, IEnum, IQueryResult, MutationSchema, TableSchema } from 'jde-framework';
 import { Instance } from './app/app.service.types';
 import * as CommonProto from '../proto/Common'; import ELogLevel = CommonProto.Jde.Proto.ELogLevel; import IException = CommonProto.Jde.Proto.IException;
-import { J } from '@angular/cdk/keycodes';
+import { AuthStore } from './auth.store';
+import { assert, Mutation } from 'jde-framework';
+import { computed, Signal } from '@angular/core';
+import { EProvider, LoggedInUser } from 'jde-material';
 
-interface IStringRequest<T>{ requestId:number; type:T; value:string; }
 interface IStringResult{ id:number; value:string; }
-interface IMessageUnion{ stringResult:IStringResult }
 export interface IError{ requestId?:number; message: string; sc?:number; }
 
 type TransformInput = (x:any)=>any;
@@ -18,13 +19,14 @@ export type RequestId = number;
 export enum ETransport{ Unsecure, Secure, Hybrid };
 
 class RequestPromise<ResultMessage>{
-	constructor( public result:(ResultMessage)=>any, public resolve:Resolve, public reject:Reject, public transformInput:TransformInput=null )
+	constructor( public result:(ResultMessage)=>any, public resolve:Resolve, public reject:Reject, public transformInput:TransformInput|null=null )
 	{}
 }
 
 export abstract class ProtoService<Transmission,ResultMessage>{
-	constructor( private TCreator: { new (): Transmission; }, protected http: HttpClient, public readonly transport:ETransport )
+	constructor( private TCreator: { new (): Transmission; }, protected http: HttpClient, public readonly transport:ETransport, protected authStore:AuthStore, private isAppServer:boolean=false )
 	{}
+
 	connect():void{
 		this.#socket = webSocket<protobuf.Buffer>( {url: this.socketUrl, deserializer: msg => this.onMessage(msg), serializer: msg=>msg, binaryType:"arraybuffer"} );
 		this.#socket.subscribe(
@@ -36,6 +38,29 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 	//should deserialize put into a constant variable or process in deserialization?
 	addMessage( msg )
 	{}
+
+	toCollectionName( collectionDisplay:string ):string{ return collectionDisplay; }
+	excludedColumns( tableName: string ): string[] { return []; }
+	subQueries( typeName: string, id: number ):string[]{ return []; }
+	targetQuery( schema: TableSchema, target: string, showDeleted:boolean ):string{
+		let fields = this.fieldColumns( schema, showDeleted );
+		return `${schema.singular}( target:"${target}" ){ ${fields.join(" ")} }`;
+	}
+	protected fieldColumns( schema: TableSchema, showDeleted:boolean ):string[]{
+		let columns = [];
+		let filtered = schema.fields.filter(
+			(x)=>!this.excludedColumns(schema.collectionName).includes(x.name) && (x.name!="deleted" || showDeleted) );
+		for( const field of filtered ){
+			if( field.type.underlyingKind==FieldKind.UNION )
+				columns.push( `${field.name}{id}` );
+			else if( field.type.underlyingKind==FieldKind.OBJECT )
+				columns.push( `${field.name}{id name}` );
+			else
+				columns.push( field.name );
+		}
+		return columns;
+	}
+
 	//
 	error( err ){
 		this.setSocketId( 0 );
@@ -65,7 +90,7 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 	}
 
 	protected async sendAuthorization( socketId:number ):Promise<void>{
-		await this.sendPromise( {sessionId:Number(`0x${this.authorization}`)}, `sendAuthorization: ${this.authorization}` );
+		await this.sendPromise( {sessionId:this.user().authorization}, `sendAuthorization: ${this.user().authorization}` );
 		this.setSocketId( socketId );//release buffer.
 	}
 
@@ -76,63 +101,114 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 		});
 	}
 
-	async InitWait():Promise<void>{
+	async initWait():Promise<void>{
 		let p = new Promise<void>( (resolve,reject)=>this.#initCallbacks.push({resolve:resolve,reject:reject}) );
 		await p;
 	}
+
+	async loginWait<Y>( target:string ):Promise<Y>{
+		let p = new Promise<Y>( (resolve,reject)=>{
+			this.#loginCallbacks.push( {target: target, resolve:resolve,reject:reject} );
+		});
+		if( this.#loginCallbacks.length==1 ){
+			let url = this.urlWithTarget( "serverSettings", true );
+			if( this.log.restRequests ) console.log( url );
+			let settings;
+			try{
+				let args = this.user()?.authorization ? {headers:{"Authorization":this.user().authorization}} : {};
+				const settings = await firstValueFrom(  this.http.get<any>(url, args) );
+				if( this.log.restResults ) console.log( JSON.stringify(settings) );
+				this.timeoutSeconds = fromIsoDuration( settings["restSessionTimeout"] );
+				let instance = parseInt( settings["serverInstance"] );
+				let active = settings["active"];
+				let timedout = this.lastRestCall && ( this.lastRestCall.getTime() < Date.now() - this.timeoutSeconds*1000 );
+				let previousInstanceIndex = this.user()?.serverInstances?.findIndex( x=>x.url==this.url ) ?? -1;
+				let previousInstance = previousInstanceIndex>=0 ? this.user().serverInstances[previousInstanceIndex].instance : 0;
+				if( !active || timedout || (this.isAppServer && instance!=previousInstance) )
+					this.authStore.reset( {url:this.url, instance:instance} );
+				else if( previousInstance!=instance )
+					this.authStore.setServerInstance( this.url, instance );
+				for( let callback of this.#loginCallbacks ){
+					let y = await this.authGet<any>( callback.target, this.user().authorization );
+					callback.resolve( y );
+				}
+			}
+			catch( e ){
+				for( let callback of this.#loginCallbacks )
+					callback.reject( e );
+			}
+			this.#loginCallbacks.length=0;
+		}
+		return p;
+	}
+
 	urlWithTarget( suffix:string, preferSecure:boolean=false ):string{
 		return preferSecure && this.transport==ETransport.Hybrid
 			? `${this.secureRestUrl}/${suffix}`
 			: `${this.restUrl}/${suffix}`;
 	}
 
-	async get<Y>( target:string ):Promise<Y>{
-		if( !this.#instances )
-			await this.InitWait();
+	private async authGet<Y>( target:string, authorization?:string ):Promise<Y>{
 		if( this.log.restRequests )	console.log( target.substring(0,this.log.maxLength) );
 		let url = this.urlWithTarget(target);
 		let y:Y;
-		try{
-			if( this.authorization )
-				y = await firstValueFrom( this.http.get<Y>(url, {headers:{"Authorization":this.authorization}}) );
-			else{
-				let response:HttpResponse<Y> = await firstValueFrom( this.http.get<Y>(url, {observe: "response", transferCache:{includeHeaders:["Authorization"]}}) );
-				let authorization = response.headers.get( "Authorization" );
-				if( authorization )
-					this.setAuthorization( authorization, null );
-				y = response.body;
+		if( authorization ){
+			try{
+				y = await firstValueFrom( this.http.get<Y>(url, {headers:{"Authorization":authorization}}) );
+			}
+			catch( e ){
+				if( e["status"]==401 ){
+					console.log( `(${e["status"]})${e["error"]} - ${e["url"]}` );
+					this.authStore.logout();
+					y = await this.authGet<Y>( target );
+				}
+				else
+					throw e;
 			}
 		}
-		catch( e ){
-			const errorResponse = e as HttpErrorResponse;
-			if( errorResponse.status==401 && this.anonymous && this.authorization ){
-				this.setAuthorization( null, null );
-				y = await this.get<Y>( target );
-			}
-			else
-				throw e;
+		else{
+			let response:HttpResponse<Y> = await firstValueFrom( this.http.get<Y>(url,
+				{observe: "response", transferCache:{includeHeaders:["Authorization"]}}) );
+			let newAuth = response.headers.get( "Authorization" );
+			if( newAuth )
+				this.authStore.append( {authorization:newAuth} );
+			y = response.body;
 		}
+		if( this.log.restResults ) console.log( JSON.stringify(y) );
+		this.lastRestCall = new Date();
 		return y;
 	}
+
+	async get<Y>( target:string ):Promise<Y>{
+		if( !this.#instances )
+			await this.initWait();
+		let isActive = this.lastRestCall && (this.lastRestCall.getTime() > Date.now() - this.timeoutSeconds*1000);
+		let y = !this.user()?.authorization || !isActive
+			? await this.loginWait<Y>( target )
+			: await this.authGet<Y>( target, this.user().authorization );
+		return y;
+	}
+
 	async postRaw<Y>( target:string, body:any, preferSecure:boolean=false ):Promise<Y>{
 		if( !this.#instances )
-			await this.InitWait();
+			await this.initWait();
 		const url = this.urlWithTarget( target, preferSecure );
 		let y:any;
-		if( this.authorization )
+		if( this.user()?.authorization )
 			y =  await firstValueFrom( this.http.post(url, body) );
 		else{
 			let response:HttpResponse<Y> = await firstValueFrom( this.http.post<Y>(url, body, {observe: "response", transferCache:{includeHeaders:["Authorization"]}}) );
 			let authorization = response.headers.get( "Authorization" );
+			console.assert( authorization!=null, "no authorization" );
 			if( authorization )
-				this.setAuthorization( authorization, null );
+				this.authStore.append( {authorization:authorization} );
 			y = response.body;
 		}
 		return y;
 	}
 	//TODO change to use postRaw
 	async post<Y>( target:string, body:any, preferSecure:boolean=false ):Promise<Y>{
-		return await this.postRaw<Y>( target, body, preferSecure )[ "value" ];
+		return (await this.postRaw<Y>( target, body, preferSecure ))[ "value" ];
 	}
 
 	private async graphQL<Y>( query: string ):Promise<Y>{
@@ -141,44 +217,85 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 		return y ? y["data"] : null;
 	}
 
-	async query<Y>( ql: string ):Promise<Y>{
-		return await this.graphQL( `query ${ql}` );
+	async providers():Promise<EProvider[]>{
+		const ql = `__type(name: "Provider") { enumValues { id name } }`;
+		const data = await this.query( ql );
+		return data["__type"]["enumValues"].map( (x:IEnum)=>x.id );
 	}
-
+	async query<Y>( ql: string ):Promise<Y>{
+		return await this.graphQL( ql );
+	}
+	async querySingle<Y>( ql: string ):Promise<Y>{
+		const y = await this.query<any>( ql );
+		return y[Object.keys(y)[0]];
+	}
+	async queryObject<Y>( ql: string, cnstrctr: new(...args:any[]) => Y ):Promise<Y>{
+		const result = await this.query<any>( ql );
+		return new cnstrctr( result[Object.keys(result)[0]] );
+	}
 	async queryArray<Y>( ql: string ):Promise<Y[]>{
 		const member = ql.substring( 0, ql.indexOf('{') ).trim();
-		const y = await this.graphQL( `query ${ql}` );
+		const y = await this.graphQL( ql );
 		return y[member];
 	}
 
-	async mutation<Y>( ql: string ):Promise<Y>{
-		return await this.graphQL( `mutation ${ql}` );
+	async querySetting(target:string):Promise<string>{
+		const queryResult = await this.querySingle<string>( `setting(target:\"${target}\"){value}` );
+		return queryResult["value"];
+	}
+	async querySettings(target:string[]):Promise<{[key:string]:string}>{
+		const queryResult = await this.query<string>( `settings(target:${JSON.stringify(target)}){target value}` );
+		let y:{[key:string]:string} = {};
+		for( const setting of queryResult["settings"] )
+			y[setting.target] = setting.value;
+		return y;
 	}
 
-	schema( names:string[] ):Promise<Table[]>{
-		return new Promise<Table[]>( (resolve, reject)=>{
-			let results = new Array<Table>();
-			let query =  new Array<string>();
-			names.forEach( (x)=>{ if( ProtoService.#tables.has(x) ) results.push(ProtoService.#tables.get(x)); else query.push(x); } );
-			if( !query.length )
-				resolve( results );
-			else{
-				for( let name of query ){
-					let ql = `__type(name: "${name}") { fields { name type { name kind ofType{name kind} } } }`;
-					this.query( ql ).then( ( data:any )=>
-					{
-						let table = new Table( data.__type );
-						ProtoService.#tables.set( name, table );
-						if( results.push( table )==names.length )
-							resolve( results );
-					}).catch( (e)=>reject(e) );
-				}
-			}
-		});
+	async mutation<Y>( ql: string|Mutation|Mutation[] ):Promise<Y>{
+		if( Array.isArray(ql) ){
+			let y = [];
+			for( let m of <Mutation[]>ql )
+				y.push( await this.mutation( m ) );
+			return <Y>y;
+		}
+		let query = ql instanceof Mutation ? ql.toString() : ql;
+		assert( query );
+		return await this.graphQL<Y>( `mutation ${query}` );
 	}
-	mutations():Promise<Mutation[]>
-	{
-		return new Promise<Mutation[]>( (resolve, reject)=>{
+
+	async schemaWithEnums( type:string ):Promise<TableSchema>{
+		let schema = ( await this.schema([type]) )[0];
+		if( !schema.enums ){
+			schema.enums = new Map<string, IEnum[]>();
+			for( const field of schema.fields.filter((x)=>x.type.underlyingKind==FieldKind.ENUM && !schema.enums.has(x.name)) ){
+				let values:Array<IEnum>;
+				values = ( await this.query<IQueryResult<IEnum>>(` __type(name: "${field.type.name}") { enumValues { id name } }`) ).__type["enumValues"];
+				schema.enums.set( field.type.name, values );
+			}
+		}
+		return schema;
+	}
+	async schema( types:string[] ):Promise<TableSchema[]>{
+		let results = new Array<TableSchema>();
+		let queries =  new Array<string>();
+		for( let type of types ){
+			if( this.#tables.has(type) )
+				results.push(this.#tables.get(type));
+			else
+				queries.push(type);
+		};
+		for( let type of queries ){
+			const ql = `__type(name: "${type}") { fields { name type { name kind ofType{name kind} } } }`;
+			const data = await this.query( ql );
+			const schema = new TableSchema( data["__type"] );
+			this.#tables.set( type, schema );
+			results.push( schema );
+		}
+		return results;
+	}
+
+	mutations():Promise<MutationSchema[]>{
+		return new Promise<MutationSchema[]>( (resolve, reject)=>{
 			if( ProtoService.#mutations )
 				resolve( ProtoService.#mutations );
 			else{
@@ -196,16 +313,6 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 	//get nextRequestId():RequestId{ return this.#requestId+1; }  why?
 	getRequestId():RequestId{ return ++this.#requestId;} #requestId:RequestId=0;
 
-	protected setAuthorization( authorization:string, loginName:string ){
-		console.log( `setAuthorization( ${authorization}, ${loginName} )` );
-		if( authorization )
-			localStorage.setItem( "authorization", authorization );
-		else
-			localStorage.removeItem( "authorization" );
-		if( loginName)
-			this.#loginNameSubscription.next( loginName );
-	}
-	subscribeLoginName():Observable<string>{ return this.#loginNameSubscription.asObservable(); }
 	protected setSocketId( id:number ){
 		this.#socketId = id;
 		for( var m of this.backlog )
@@ -257,11 +364,8 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 	protected abstract handleConnectionError( err );
 	protected abstract encode( t:Transmission );
 
-	private anonymous:boolean=true;
 	protected backlog:Transmission[] = [];
-	protected log = { sockRequests:true, sockResults:true, restRequests:true, restResults:false, subRequest:true, subResults:true, maxLength:255 };
-	#loginNameSubscription:Subject<string> = new Subject<string>();
-	protected get authorization():string{ return localStorage.getItem("authorization"); }
+	protected log = { sockRequests:true, sockResults:true, restRequests:false, restResults:false, subRequest:true, subResults:true, maxLength:255 };
 	//Informational purposes only to match with server logs.
 	protected get socketId():number{ return this.#socketId; } #socketId:number;
 	get instances(){return this.#instances;} set instances(x){
@@ -274,11 +378,13 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 		}
 	} #instances:Instance[];
 	#initCallbacks:{resolve:()=>void, reject:Reject}[]=[];
+	#loginCallbacks:{target:string, resolve:(result:any)=>void, reject:( e:any )=>void}[]=[];
+
 	//abstract get queryId():number;
 	#socket:WebSocketSubject<protobuf.Buffer>;
 	protected _callbacks = new Map<number, RequestPromise<ResultMessage>>();
-	static #tables = new Map<string,Table>();
-	static #mutations:Array<Mutation>;
+	#tables = new Map<string,TableSchema>();
+	static #mutations:Array<MutationSchema>;
 	private get url(){
 		if( !this.instances?.length ) throw "no instances";
 		return `${this.instances[0].host}:${this.instances[0].port}`;
@@ -286,4 +392,9 @@ export abstract class ProtoService<Transmission,ResultMessage>{
 	protected get socketUrl(){ return `${this.transport==ETransport.Secure ? "wss" : "ws"}://${this.url}`; }
 	private get restUrl(){return this.transport==ETransport.Secure ? this.secureRestUrl : `http://${this.url}`;}
 	private get secureRestUrl(){return `https://${this.url}`;}
+
+	isLoggedIn = computed( () => this.user()!= null );
+	get user():Signal<LoggedInUser>{return this.authStore.user; }
+	lastRestCall:Date = null;
+	timeoutSeconds:number;
 }
