@@ -1,15 +1,18 @@
 import {Component, Input, Output, OnInit, EventEmitter, NgModule,ViewChild} from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import {MatChipsModule} from '@angular/material/chips';
-import {MatDatepickerInputEvent, MatDatepickerModule} from '@angular/material/datepicker';
-import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatChipListbox, MatChipOption, MatChipsModule} from '@angular/material/chips';
+import {MatDatepickerInputEvent, MatDatepickerModule, MatDatepickerToggle, MatDateRangeInput, MatDateRangePicker} from '@angular/material/datepicker';
+import {MatFormField, MatFormFieldModule, MatLabel} from '@angular/material/form-field';
 
 import { DateUtilities, Day } from '../../utilities/dateUtilities';
 import { IAssignable } from '../../utilities/settings';
 
 export enum TimeFrame{None=0, Week=7, Month=30, Quarter=90, Year=360, All=1000}
 
-@Component( {selector: 'date-range',templateUrl: 'date-range.html'} )
+@Component( {
+	selector: 'date-range',
+	templateUrl: 'date-range.html',
+	imports: [MatFormField, MatLabel, MatDateRangeInput, MatDatepickerToggle, MatDateRangePicker,MatChipListbox,MatChipOption]} )
 export class DateRange implements OnInit
 {
 	get startDate()
@@ -72,7 +75,7 @@ export class DateRange implements OnInit
 	{
 		return DateUtilities.toDays( new Date(time.getTime()-time.getTimezoneOffset()*60000) );
 	}
-	startChange( e:MatDatepickerInputEvent<Date> )
+	startChange( e:MatDatepickerInputEvent<Date>|any )
 	{
 		console.log( `startChange = ${e.value}  - start=${this.settings.start} - end=${this.settings.end ?? this.settings.max}, days=${this.settings.dayCount}` );
 		this.settings.timeFrame = null;
@@ -84,8 +87,7 @@ export class DateRange implements OnInit
 //		console.log( 'start='+this.range.controls['start'].value );
 //		console.log( 'end='+this.range.controls['end'].value );
 	}
-	endChange( e:MatDatepickerInputEvent<Date> )
-	{
+	endChange( e:MatDatepickerInputEvent<Date>|any ){
 		console.log( `endChange = ${e.value}  - start=${this.settings.start} - end=${this.settings.end ?? this.settings.max}, days=${this.settings.dayCount}` );
 		this.settings.end = e.value ? DateUtilities.toDays( e.value ) : null;
 		//console.log( 'start='+this.range.controls['start'].value );
@@ -153,7 +155,8 @@ export class DateRangeSettings implements IAssignable<DateRangeSettings>
 
 	get start(){ return this.#start; } set start( x:Day ){ this.#start = x; }  #start:Day;
 }
-
+/*
 @NgModule( {exports: [DateRange], declarations: [DateRange], imports:[MatChipsModule,MatDatepickerModule,MatFormFieldModule]} )
 export class DateRangeModule
 {}
+*/
